# Item browser performance improvements