--- a/src/wwwroot/css/genpage.css
+++ b/src/wwwroot/css/genpage.css
@@ -252,6 +252,8 @@ body {
 .image-block img, .img-block video {
     display: block;
     margin: auto;
+    padding-left: 0.4rem;
+    padding-right: 0.4rem;
 }
 .image-block-placeholder img {
     filter: blur(5px);
@@ -262,7 +264,6 @@ body {
 .image-block-legacy {
     position: relative;
     height: 11.5rem;
-    width: 10rem;
     padding-top: 0.1rem;
     padding-left: 0.1rem;
     padding-right: 0.1rem;
@@ -272,15 +273,12 @@ body {
 }
 .image-block-small {
     height: 7rem;
-    width: 7rem;
 }
 .image-block-big {
     height: 17.5rem;
-    width: 16rem;
 }
 .image-block-giant {
     height: 27.5rem;
-    width: 26rem;
 }
 .image-block-legacy img {
     height: 8rem;
@@ -366,6 +364,9 @@ body {
 }
 .image-preview-text {
     font-family: var(--font-monospace);
+    /* this dark magic will make the text grow as big as the parent container is and then wrap (instead of growing the container even more) */
+    min-width: 100%;
+    max-width: 0;
 }
 .current_model_view {
     display: inline-block;