@brendanhoar 2025-07-20
I'm excited!

But also, I don't really count here, lol.